pyproject.toml
[tool.poetry]
name = "claim-reports"
version = "0.1.0"
description = "drf application for creating claim reports"
authors = ["Alexander Dovguchits <alexander.dovguchits@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
django-rest-knox = "^4.2.0"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.8.0"
django-cors-headers = "^4.3.1"
django-cron = "^0.6.0"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.1.1"
black = "^23.3.0"
django-stubs = "^4.2.0"
djangorestframework-stubs = "^3.14.0"
flake8 = "^6.0.0"
flake8-annotations-complexity = "^0.0.7"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.5.9"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.12.0"
flake8-debugger = "^4.1.2"
flake8-executable = "^2.1.3"
flake8-expression-complexity = "^0.0.11"
flake8-fixme = "^1.1.1"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.0.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-quotes = "^3.3.2"
flake8-simplify = "^0.20.0"
flake8-variables-names = "^0.0.5"
httpie = "^3.2.2"
ipython = "^8.13.2"
mypy = "^1.3.0"
pep8-naming = "^0.13.3"
pyproject-flake8 = "^6.0.0.post1"
types-requests = "^2.30.0.0"
pytest = "^7.3.1"
jsonschema = "^4.17.3"
flake8-import-order = "^0.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
